requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.25.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
bcrypt>=4.0.1
pyjwt>=2.8.0
python-multipart>=0.0.6

